moved findSelectedCatalogueItemInSession from additional-service controller into helper\routes
