+++ app/pages/sections/order-items/additional-services/select/routes.test.js
@@ -18,9 +18,11 @@ import { routes } from '../../../../../routes';
 import { baseUrl } from '../../../../../config';
 import { getRecipients } from '../../../../../helpers/api/ordapi/getRecipients';
 import * as routerHelper from '../../../../../helpers/routes/routerHelper';
+import { findSelectedCatalogueItemInSession } from '../../../../../helpers/routes/findSelectedCatalogueItemInSession';
 
 jest.mock('../../../../../logger');
 jest.mock('../../../../../helpers/api/ordapi/getRecipients');
+jest.mock('../../../../../helpers/routes/findSelectedCatalogueItemInSession');
 
 const mockLogoutMethod = jest.fn().mockResolvedValue({});
 
@@ -253,8 +255,7 @@ describe('additional-services select routes', () => {
       selectAdditionalServiceController.findAdditionalServices = jest.fn()
         .mockResolvedValue(additionalServices);
 
-      selectAdditionalServiceController.findSelectedCatalogueItemInSession = jest.fn()
-        .mockResolvedValue({ name: 'Additional Service 1' });
+      findSelectedCatalogueItemInSession.mockResolvedValue({ name: 'Additional Service 1' });
 
       selectAdditionalServiceController.getAdditionalServicePageContext = jest.fn()
         .mockResolvedValue({});
